done Sprint #9 v1.1; fix: изменить проверку тестов, валидацию film. style: убрать весь неиспользованный код
