+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Data;
 import ru.yandex.practicum.filmorate.validator.MovieBirthday;
 
-import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
@@ -13,12 +12,14 @@ import java.time.LocalDate;
 @Data
 @Builder
 public class Film {
+    final String FIRST_RELEASE = "1895-12-28";
+
     private Long id;
     @NotBlank(message = "Название фильма не может быть пустым")
     private String name;
     @Size(max = 200, message = "Максимальная длина описания более 200 символов")
     private String description;
-    @MovieBirthday(message = "Дата релиза раньше 28 декабря 1895 года")
+    @MovieBirthday(value = FIRST_RELEASE)
     private LocalDate releaseDate;
     @Min(value = 1, message = "Продолжительность фильма нулевая или отрицательная")
     private long duration;
