+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -9,13 +9,8 @@ import java.time.LocalDate;
 @Builder
 public class User {
     private Long id;
-//    @Email(message = "Не корректный адрес электронной почты")
-//    @NotBlank(message = "Поле email не может быть пустым")
     private String email;
-//    @NotBlank(message = "Поле Login не может быть пустым")
     private String login;
     private String name;
-//    @Past(message = "День рождения не может быть в будущем")
-//    @NotNull(message = "День рождение не может быть пустым")
     private LocalDate birthday;
 }
