+++ src/main/java/ru/yandex/practicum/filmorate/validator/MovieBirthday.java
@@ -1,24 +1,21 @@
 package ru.yandex.practicum.filmorate.validator;
 
-import ru.yandex.practicum.filmorate.validator.MovieBirthdayValidator;
-
 import javax.validation.Constraint;
 import javax.validation.Payload;
-import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target({FIELD})
-@Retention(RUNTIME)
 @Constraint(validatedBy = MovieBirthdayValidator.class)
-@Documented
+@Target({ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
 public @interface MovieBirthday {
-    String message() default "{CapitalLetter.invalid}";
+    String message() default "Дата релиза должна быть позже {value}";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
+
+    String value();
 }
